feat(db): setting up basic sqlite image
Using docker-compose. The image used ('nouchka/sqlite3') is not
suited for production as the image 'keinos/sqlite3', but i had
problems setting that one up, as the documentation is not
very clear on how to navigate all the security configs, and
this is just a simple golang exercise.

Some sources used:
www.mustafabugra.com/development/sqlite-3-docker-compose
database.guide/5-ways-to-run-sql-script-from-file-sqlite